Add i2c_{read_from,write_to}
These functions implement the pattern where first a value is written
that holds the register that the subsequent bytes should be read from or
written to. For reads, this means doing a write followed by a read. For
writes, this means doing a longer write where the payload is prefixed by
the first register it should be written to.