Remove need for idle task
Prior to this change, every context switch event would check if there is
another task that can be run, and if not the case, would switch to
running the idle task. The idle task was treated like any other task,
with its register context saved and restored upon exit and entry.
Additionally, it would occupy the same amount of stack space as other
(more useful) tasks.

This change removes the need for an idle task. Instead, every context
switch will jump to a scheduling function, which will either switch to
another runnable task, or go to sleep. Where prior to this change there
would _always_ be a running task (real or idle), this is no longer the
case. When sleeping in the scheduling function, the current task pointer
is now NULL. Since the task timer interrupt can be raised at any time,
the function to save a task's context needs an extra check to see if
there is a current task to begin with. If there is not, this function
doesn't save any context and returns quickly. This check is added as
assembly code in the `task__push` function, since this is where the cost
of doing the check is lowest (in the best case scenario, where is a task
context to save, the extra checks add 4 cycles over the previous code).

By breaking the symmetry of always saving a task's context and restoring
another (or the same) task's context, it is now possible to re-enable
interrupts for a brief duration after saving context and before
restoring context. During this time, ISRs (besides the task timer
interrupt) can be handled on the stack allocated for the scheduling
function. This is, however, not implemented in this commit.